\name{calculateCoeffs.polr}
\alias{calculateCoeffs.polr}
\title{Calculate coefficients and standard errors based on weights only for polr model}
\usage{
  calculateCoeffs.polr(model, datas, method="probit", weigths, fweigth)
}
\description{
  This function run polr model for diifferent weighting schemas, and then calculate the overall coefficients and standard errors
}
\references{
}
\author{
 Przemyslaw Biecek
}
\examples{
  
}
\keyword{datasets}
