\name{countryCentroids}
\alias{countryCentroids}
\docType{data}
\title{Country centroids}
\description{
Dataset with coordinate of country centroids

Name - Country name
ISO2 - Country ISO2 code
ISO3 - Country ISO3 code
latitude, longitude - coordinates

}
\usage{data(countryCentroids)}
\author{
 Dataset created by: CIA World Factbook, transformed by Przemyslaw Biecek
}
\examples{
head(countryCentroids)
}
\keyword{datasets}
