\name{ISCO08}
\alias{ISCO08}
\docType{data}
\title{ISCO 08 - International Standard Classification of Occupations 2008}
\description{
 Occupations in PISA are described with the use of ISCO 08 classification (for PISA 2012).

 Details are in the pdf format in codebooks.
 Here there are in R format to make it easier to use.

 Details of these classification can be found here:
 \link{\code{http://www.ilo.org/public/english/bureau/stat/isco/intro.htm}}
}
\usage{data(ISCO08)}
\author{
 Dataset created by ILO, transformed by Przemyslaw Biecek
}
\examples{
head(ISCO08)
}
\keyword{datasets}
