\name{ISCO88}
\alias{ISCO88}
\docType{data}
\title{Classification of Occupations in PISA}
\description{
 Occupations in PISA are described with the use of ISCO 88 classification.

 Details are in the pdf format in codebooks.
 Here there are in R format to make it easier to use.

 Details of these classification can be found here:
 \link{\code{http://www.ilo.org/public/english/bureau/stat/isco/intro.htm}}
}
\usage{data(ISCO88)}
\author{
 Dataset created by ILO, transformed by Przemyslaw Biecek
}
\examples{
head(ISCO88)
}
\keyword{datasets}
