\name{ISEIconv}
\alias{ISEIconv}
\docType{data}
\title{Converion table from ISCO-88 to ISEI}
\description{
 ISEI stands for International Socio-economic Index of Occupational Status. It can be derived out of ISCO 88 classification.

 Details of the translation can be found here:
 \link{\code{http://home.fsw.vu.nl/hbg.ganzeboom/pdf/1992-ganzeboom-degraaf-treiman-isei68-(ssr).pdf}} (article)
 and here 
 \link{\code{http://home.fsw.vu.nl/hbg.ganzeboom/isco88/index.htm}} (website)
}
\usage{data(ISEIconv)}
\author{
 Harry Ganzeboom (author of the conversion table)
}
\examples{
head(ISEIconv)
}
\keyword{datasets}
