\name{ISEIconv08}
\alias{ISEIconv08}
\docType{data}
\title{Converion table from ISCO-808 to ISEI}
\description{
 ISEI stands for International Socio-economic Index of Occupational Status. It can be derived out of ISCO 08 classification.

 }
\usage{data(ISEIconv08)}
\examples{
head(ISEIconv08)
}
\keyword{datasets}
