\name{computerMathClassification}
\alias{computerMathClassification}
\alias{computerProblemSolvingClassification}
\alias{computerReadingClassification}
\alias{paperFinanceClassification}
\alias{paperMathClassification}
\alias{paperReadingClassification}
\alias{paperScienceClassification}

\docType{data}
\title{Various classification of Math, Reading, Science, Financial Items}
\description{
 The official classifications for computer and paper, math reading finance and problem solving items.

 Different domains have different classifications.

 The common one is the item format.
}
\usage{data(computerMathClassification)}
\author{
 Dataset created by OECD, transformed by Przemyslaw Biecek
}
\examples{
head(computerMathClassification)
head(computerProblemSolvingClassification)
head(computerReadingClassification)
head(paperFinanceClassification)
head(paperMathClassification)
head(paperReadingClassification)
head(paperScienceClassification)
}
\keyword{datasets}
