\name{plotMeltedItemGroupPerformance}
\alias{plotMeltedItemGroupPerformance}
\alias{meltedItemGroupPerformance}
\alias{itemGroupPerformance}
\title{Plot of item performance in groups}
\description{
 Set of functions that presents item performance in groups.
}
\usage{
 itemGroupPerformance(itemPerformance, itemClassification, allItemsName = " Average Improvement Across All Items")

 meltedItemGroupPerformance(igPerformance) 

 plotMeltedItemGroupPerformance(migPerformance, selectedCnt = "F011") 
}
\author{
 Przemyslaw Biecek
}
\examples{
 # will be avaiable in 2014 when PISA2012lite be public
}
\keyword{datasets}
