\name{mathItemClassification}
\alias{mathItemClassification}
\docType{data}
\title{Classification of Math items}
\description{
 The oroginal and the ,,extended'' classification of math items.

 Each item is classified into a set of different scales.
  
 "Content Area" "Overarching idea" "Competency clusters" "Item Format" "Thinking across multiple representations" "Probabilistic reasoning" "Statistics" "Rates Ratios Proportions" "Measurement" "Using a graph chart or table" "Using multiple graphs" "Spatial reasoning" "Conceptual understanding vs procedural knowledge" "Argument explanation" "Systematic sequential strategic" "Studying Examples" "Thinking through a multi step solution process" "Employing conceptual understanding reasoning" "Using reverse thinking flexible thinking" "Making sense of non routine situations" 
}
\usage{data(mathItemClassification)}
\author{
 Dataset created by OECD, transformed by Przemyslaw Biecek
}
\seealso{\code{\link{student2009}}}
\examples{
head(mathItemClassification)
}
\keyword{datasets}
